Move the seeders into test directory
Also renamed the DatabaseSeeder to OAuth2DatabaseSeeder.

This closes #471, #467 and #435
